requirements.txt
asgiref==3.2.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-cors-headers==3.2.0
django-environ==0.4.5
djangorestframework==3.11.0
django-rest-auth==0.9.5
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
urllib3==1.25.7
